ML/AWS-ML-test2/app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
pillow==9.3.01
